Implementiran middleware.

+++ BookstoreApplication/BookstoreApplication/Services/BookService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using BookstoreApplication.DTOs;
+using BookstoreApplication.Exceptions;
 using BookstoreApplication.Interfaces;
 using BookstoreApplication.Models;
 
@@ -26,40 +27,52 @@ namespace BookstoreApplication.Services
             return books.Select(_mapper.Map<BookDto>).ToList();
         }
 
-        public async Task<BookDetailsDto?> GetByIdAsync(int id)
+        public async Task<BookDetailsDto> GetByIdAsync(int id)
         {
             var book = await _bookRepository.GetByIdAsync(id);
-            if (book == null) return null;
+            if (book == null)
+                throw new NotFoundException(id);
             return _mapper.Map<BookDetailsDto>(book);
         }
 
-        public async Task<Book?> AddAsync(Book book)
+        public async Task<Book> AddAsync(Book book)
         {
             var author = await _authorRepository.GetByIdAsync(book.AuthorId);
-            if (author == null) return null;
+            if (author == null)
+                throw new BadRequestException($"Author with id {book.AuthorId} does not exist.");
 
             var publisher = await _publisherRepository.GetByIdAsync(book.PublisherId);
-            if (publisher == null) return null;
+            if (publisher == null)
+                throw new BadRequestException($"Publisher with id {book.PublisherId} does not exist.");
 
             return await _bookRepository.AddAsync(book);
         }
 
-        public async Task<Book?> UpdateAsync(int id, Book book)
+        public async Task<Book> UpdateAsync(int id, Book book)
         {
+            if (id != book.Id)
+                throw new BadRequestException("Identifier value is invalid.");
+
             var existing = await _bookRepository.GetByIdAsync(id);
-            if (existing == null) return null;
+            if (existing == null)
+                throw new NotFoundException(id);
 
             var author = await _authorRepository.GetByIdAsync(book.AuthorId);
-            if (author == null) return null;
+            if (author == null)
+                throw new BadRequestException($"Author with id {book.AuthorId} does not exist.");
 
             var publisher = await _publisherRepository.GetByIdAsync(book.PublisherId);
-            if (publisher == null) return null;
+            if (publisher == null)
+                throw new BadRequestException($"Publisher with id {book.PublisherId} does not exist.");
 
             return await _bookRepository.UpdateAsync(book);
         }
 
         public async Task<bool> DeleteAsync(int id)
         {
+            var exists = await _bookRepository.GetByIdAsync(id);
+            if (exists == null)
+                throw new NotFoundException(id);
             return await _bookRepository.DeleteAsync(id);
         }
     }
